requirements.txt
selenium
pydantic
python-dotenv
pyyaml
